% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoevodynamo-package.R
\docType{package}
\name{ecoevodynamo-package}
\alias{ecoevodynamo}
\alias{ecoevodynamo-package}
\title{ecoevodynamo: Ecoevolutionary Dynamics Simulation Using Automatic Differentiation}
\description{
The goal of {ecoevodynamo} is to make it easy to generate ecoevolutionary 
    dynamics simulations when you only have equations for the ecological dynamics, by
    using automatic differentiation based on the {torch} package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rdinnager.github.io/ecoevodynamo}
}

}
\author{
\strong{Maintainer}: Russell Dinnage \email{r.dinnage@gmail.com} (\href{https://orcid.org/0000-0003-0846-2819}{ORCID}) [copyright holder]

}
\keyword{internal}
