% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{eed_ecodynamo}
\alias{eed_ecodynamo}
\title{Create an ecodynamo object for running an ecological dynamics ODE system.}
\usage{
eed_ecodynamo(eco, example_inputs, jit = FALSE, gpu = FALSE)
}
\arguments{
\item{eco}{An expression (see \code{\link[=eed_eco]{eed_eco()}}), a function, or an \code{\link[=eed_eco]{eed_eco()}} object
specifying ecological dynamics as a set of ordinary differential equations.
See details for instructions on how to construct this argument.}

\item{example_inputs}{A named list where the names correspond to function arguments
of the function specified in \code{eco} and the elements are example inputs for those
arguments that have the dimensions of the desired system (these could be objects you
were using for testing the \code{eco} function, or vectors, matrices or arrays filled with
\code{NA}s, just make sure they have the correct dimensions, e.g. the first dimension will
generally have a size equal to the number of species you want to simulate for the
\code{N} and the \code{X} arguments -- see details). Make sure the \code{example_inputs}
actually do work because this is necessary to set up the model correctly. If
\code{example_inputs} is \code{NULL}, the example inputs will be inferred from the global
environment if possible. If you are unsure what inputs you \code{\link[=eed_eco]{eed_eco()}} object needs,
call \code{\link[=eed_inputs]{eed_inputs()}} on it for a reminder.}

\item{jit}{Should the ecological dynamics be optimized using Just in Time (JIT) compilation?
This can speed up computations, but is currently somewhat experimental in \code{torch} so may
be unstable.}

\item{gpu}{Should the computations be run on a GPU? This again can speed up computations
for some kinds of models but adds additional overhead which will make other types of
models slower. Only does anything if you have a GPU version of \code{torch} installed (and you
have a GPU of course).}
}
\value{
An \code{ecodynamo} object which can be used to run an ecological dynamics simulation.
}
\description{
Create an ecodynamo object for running an ecological dynamics ODE system.
}
