% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{print.eed_eco}
\alias{print.eed_eco}
\title{Print an \code{eed_eco} object prettily}
\usage{
\method{print}{eed_eco}(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Print an \code{eed_eco} object prettily
}
